# api/openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadUrl": {
        "description": "A presigned GET for the playback file, tagged with the revision and start it was signed for.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "playback_rev": {
            "title": "Playback Rev",
            "type": "string"
          },
          "playback_start_ms": {
            "title": "Playback Start Ms",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at",
          "playback_rev",
          "playback_start_ms"
        ],
        "title": "DownloadUrl",
        "type": "object"
      },
...
      "PeaksUrl": {
        "description": "A presigned GET for the waveform file, tagged with the revision it was signed for.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "peaks_rev": {
            "title": "Peaks Rev",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at",
          "peaks_rev"
        ],
        "title": "PeaksUrl",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem details body, the shape of every error this API returns.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "`about:blank`, or a problem a client branches on: `urn:crosstune:account-deleted` (401, the account was deleted, so the client drops its local data), `urn:crosstune:premium-required` (403, the addition needs Premium), `urn:crosstune:scan-tune-only` (403, a free account's scans stay on one tune), `urn:crosstune:quota-exceeded` (413), `urn:crosstune:file-too-large` (413).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SignedUrl": {
        "description": "A presigned URL and when it stops working.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at"
        ],
        "title": "SignedUrl",
        "type": "object"
      },
...
      "UploadSlotRequest": {
        "description": "What the client is about to upload.",
        "properties": {
          "bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Bytes",
            "type": "integer"
          },
          "content_type": {
            "maxLength": 100,
            "pattern": "^audio/[A-Za-z0-9.+_-]+(;[A-Za-z0-9.+_=\\- ]+)*$",
            "title": "Content Type",
            "type": "string"
          }
        },
        "required": [
          "bytes",
          "content_type"
        ],
        "title": "UploadSlotRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Crosstune API",
    "version": "0.19.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/recordings/{recording_id}/download": {
      "get": {
        "description": "A presigned GET for the playback file of a ready recording.\n\nCarries the revision and start the signature was issued for, read from the same row\nas the key: a trim landing between this response and the client's GET changes the\nrow, but never what this response already promised.",
        "operationId": "download_v1_recordings__recording_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Download",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}/peaks": {
      "get": {
        "description": "A presigned GET for the waveform peaks file of a recording.\n\nCarries the revision the signature was issued for, read from the same row as the key.",
        "operationId": "peaks_v1_recordings__recording_id__peaks_get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeaksUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Peaks",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}/retry": {
      "post": {
        "description": "Process a failed recording again: transcode its file, or fetch an import again.\n\nAn import whose file never arrived is fetched again, or fails at once when its\naddress is not one the server imports from. Any other recording whose uploaded\nobject is gone is uploaded again through a new slot; this route only re-runs the\ntranscode. Fetching an import again needs Premium; a transcode does not. Repeating\nthe call changes nothing.",
        "operationId": "retry_v1_recordings__recording_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Retry",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}/upload-slot": {
      "post": {
        "description": "A presigned PUT for one recording's file, once the plan and its quota allow it.\n\nA failed recording is issued a slot too, and returns to pending_upload: it is\nthe only way back for one whose uploaded object is no longer in the bucket.",
        "operationId": "upload_slot_v1_recordings__recording_id__upload_slot_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadSlotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Upload Slot",
        "tags": [
          "recordings"
        ]
      }
    },
    "/v1/recordings/{recording_id}/uploaded": {
      "post": {
        "description": "Confirm the object landed and queue its transcode. Repeating the call changes nothing.",
        "operationId": "upload_finished_v1_recordings__recording_id__uploaded_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Upload Finished",
        "tags": [
          "recordings"
        ]
      }
    },
...
  }
}